+++ app/src/main/java/com/nimkat/app/repository/AuthRepository.kt
@@ -21,13 +21,13 @@ class AuthRepository @Inject constructor(
         return gson.fromJson(authString, AuthModel::class.java)
     }
 
-
     fun initCode(): GetCodeResponse?{
         val authString = authPrefs.getCode()
         if (authString === null || authString == "") return null
         val gson = Gson()
         return gson.fromJson(authString, GetCodeResponse::class.java)
     }
+
 
     suspend fun getCode(phoneNumber: String): Response<GetCodeResponse> {
         var apiResponse = api.getCode(GetCodeBody(phoneNumber, "-"))
@@ -50,6 +50,21 @@ class AuthRepository @Inject constructor(
         authPrefs.clearAuth()
     }
 
+
+
+
+
+
+
+//    profile Repository
+
+    fun initProfile(): ProfileModel? {
+        val profileString = authPrefs.getProfileString()
+        if (profileString === null) return null
+        val gson = Gson()
+        return gson.fromJson(profileString, ProfileModel::class.java)
+    }
+
     suspend fun updateProfile(name: String , gradeID :Int  , id: String , token: String , username: String): Response<ProfileModel>?{
 
         val apiResponse = api.updateProfile(id , ProfileInfo(name = name , grade = gradeID , username = username) ,
@@ -61,6 +76,21 @@ class AuthRepository @Inject constructor(
         authPrefs.setProfileString(gson.toJson(apiResponse.body()))
         Log.d("Auth" , "json is: " + gson.toJson(apiResponse.body()))
         return apiResponse
-
     }
+
+
+    suspend fun getProfile(id: String , token: String): Response<ProfileModel>?{
+        val apiResponse = api.getProfile(id , "Token $token")
+        if (apiResponse.body() === null) return null;
+        Log.d("Auth" , "profile status " + apiResponse.body()!!.isProfileCompleted)
+        val gson = Gson()
+        authPrefs.setProfileString(gson.toJson(apiResponse.body()))
+        Log.d("Auth" , "json is: " + gson.toJson(apiResponse.body()))
+        return apiResponse
+    }
+
+    suspend fun delete(toString: String, token: String?) {
+        api.deleteAccount(toString , "Token $token")
+    }
+
 }
